p0-money-integrity: server-side order re-pricing + identity wallet BigDecimal
Order (roadmap #3, never shipped): OrderService.createOrder now re-prices
every line from the restaurant menu snapshot via the existing cached batch
resolver (RestaurantPricedItemResolver, GET /api/v1/menu/items?ids=) BEFORE
the write tx. Client-supplied unitPrice/name can no longer reach the order
entity, saga payload, payment charge, or outbox events. Fail-closed: ids the
restaurant omits (nonexistent/unavailable/inactive) reject with the resolver's
existing BusinessException; outages surface as UpstreamUnavailableException.
All create surfaces get parity for free (they funnel through createOrder):
OrderController, CustomerOrderController, LegacyOrderCompatController
(/create, cart checkout, ?async=true job, /create-batch), reorder, assist.

Identity: customers.wallet_balance leaves binary floating point — entity +
CustomerResponse/CustomerProfileResponse Double -> BigDecimal
(precision 12, scale 2), CustomerWalletSyncPort/Adapter take BigDecimal with
a non-negative guard, additive migration
V11__customers_wallet_balance_numeric.sql (double precision -> numeric(12,2)).
Payment-side sync source (wallet_balances.balance) is already BigDecimal; no
payment edits required.

Tests: tamper (client 999.00 vs menu 240.00 -> stored at server price),
unavailable-item rejection with no persistence, server-priced reservation
lines; PG tests stub the new menu batch call. AsyncSagaRecovery + OrderSagaEventContract
received ONLY the required getMenuItems stubs (approved); their remaining 3
failures are pre-existing at base 2381ee3 (verified against a pristine base
worktree; the owning batch's uncommitted edits are their fixes), as is the
payment PaymentIdempotentReplay decline test failure (zero payment files touched).

+++ services/identity/src/main/java/com/bhukkad/identity/api/CustomerProfileResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -22,7 +23,7 @@ public class CustomerProfileResponse {
     private Boolean phoneVerified;
     private Boolean profileCompleted;
     private Integer loyaltyPoints;
-    private Double walletBalance;
+    private BigDecimal walletBalance;
     private String role;
     private String createdAt;
     private List<AddressResponse> addresses;
